LedgerMetrics using hooks and typescript
